[Feat] support float data_factor/render_factor (images_2p5).
Make data_factor float end-to-end (configs, ColmapDataParser, viewer).

Adopt images_<suffix> naming: integers keep images_2, non-integers use images_2p5/images_2p51.

Update MoGe prior inference --factor to float and align naming/messages.

Update strategies/pruning visual benchmarks to propagate float data_factor and infer from images_* dirs.

Make TSDF meshing --render_factor float.

+++ friendly_splat/trainer/configs.py
@@ -31,7 +31,7 @@ class IOConfig:
 @dataclass(frozen=True)
 class DataConfig:
     # Downsample factor for the dataset images.
-    data_factor: int = 1
+    data_factor: float = 1.0
     # Normalize the world space based on COLMAP points/cameras.
     normalize_world_space: bool = True
     # Whether to align world axes during normalization (z-up + PCA principal axes).
